api/requirements_local.txt
django-extensions==3.2.*
